add some info for the 0x209 entry
It is not known yet what this entry really is used for but it is a record_id that points to a model in the same database. Upon loading, both models are interconnected. Maybe the other value indicates some distance float value for a LOD or alternate model. This has to be checked in the future.